# scripts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BI Platform API",
    "description": "Корпоративная система бизнес-аналитики — API для ввода данных, авторизации и метрик",
    "version": "0.1.0"
  },
  "paths": {
...
    "/input/expenses": {
      "post": {
        "tags": [
          "input"
        ],
        "summary": "Create Expense",
        "operationId": "create_expense_input_expenses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/input/transactions": {
      "post": {
        "tags": [
          "input"
        ],
        "summary": "Create Transaction",
        "operationId": "create_transaction_input_transactions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/input/submit": {
      "post": {
        "tags": [
          "input"
        ],
        "summary": "Submit Entry",
        "operationId": "submit_entry_input_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataInputSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataInputOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/input/entries": {
      "get": {
        "tags": [
          "input"
        ],
        "summary": "List Entries",
        "operationId": "list_entries_input_entries_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataInputOut"
                  },
                  "title": "Response List Entries Input Entries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataInputOut": {
        "properties": {
          "entry_id": {
            "type": "integer",
            "title": "Entry Id"
          },
          "template_id": {
            "type": "integer",
            "title": "Template Id"
          },
          "submitted_by": {
            "type": "string",
            "title": "Submitted By"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "entry_data": {
            "type": "object",
            "title": "Entry Data"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "applied_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          }
        },
        "type": "object",
        "required": [
          "entry_id",
          "template_id",
          "submitted_by",
          "submitted_at",
          "entry_data",
          "status"
        ],
        "title": "DataInputOut"
      },
      "DataInputSubmit": {
        "properties": {
          "template_code": {
            "type": "string",
            "title": "Template Code"
          },
          "entry_data": {
            "type": "object",
            "title": "Entry Data"
          }
        },
        "type": "object",
        "required": [
          "template_code",
          "entry_data"
        ],
        "title": "DataInputSubmit"
      },
      "ExpenseCreate": {
        "properties": {
          "expense_date": {
            "type": "string",
            "format": "date",
            "title": "Expense Date"
          },
          "account_id": {
            "type": "integer",
            "title": "Account Id"
          },
          "department_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Id"
          },
          "employee_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Id"
          },
          "expense_category": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Expense Category"
          },
          "expense_item": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Expense Item"
          },
          "currency_id": {
            "type": "integer",
            "title": "Currency Id"
          },
          "amount_original": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Amount Original"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "receipt_info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Receipt Info"
          }
        },
        "type": "object",
        "required": [
          "expense_date",
          "account_id",
          "expense_category",
          "expense_item",
          "currency_id",
          "amount_original"
        ],
        "title": "ExpenseCreate"
      },
      "ExpenseOut": {
        "properties": {
          "expense_id": {
            "type": "integer",
            "title": "Expense Id"
          },
          "expense_date": {
            "type": "string",
            "format": "date",
            "title": "Expense Date"
          },
          "account_id": {
            "type": "integer",
            "title": "Account Id"
          },
          "amount_original": {
            "type": "string",
            "title": "Amount Original"
          },
          "amount_rub": {
            "type": "string",
            "title": "Amount Rub"
          },
          "expense_category": {
            "type": "string",
            "title": "Expense Category"
          },
          "expense_item": {
            "type": "string",
            "title": "Expense Item"
          },
          "is_manual_entry": {
            "type": "boolean",
            "title": "Is Manual Entry"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "expense_id",
          "expense_date",
          "account_id",
          "amount_original",
          "amount_rub",
          "expense_category",
          "expense_item",
          "is_manual_entry",
          "created_at"
        ],
        "title": "ExpenseOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TransactionCreate": {
        "properties": {
          "transaction_date": {
            "type": "string",
            "format": "date",
            "title": "Transaction Date"
          },
          "posting_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posting Date"
          },
          "account_id": {
            "type": "integer",
            "title": "Account Id"
          },
          "counterparty_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counterparty Id"
          },
          "department_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Id"
          },
          "currency_id": {
            "type": "integer",
            "title": "Currency Id"
          },
          "amount_original": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Amount Original"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "transaction_date",
          "account_id",
          "currency_id",
          "amount_original"
        ],
        "title": "TransactionCreate"
      },
      "TransactionOut": {
        "properties": {
          "transaction_id": {
            "type": "integer",
            "title": "Transaction Id"
          },
          "transaction_date": {
            "type": "string",
            "format": "date",
            "title": "Transaction Date"
          },
          "account_id": {
            "type": "integer",
            "title": "Account Id"
          },
          "amount_original": {
            "type": "string",
            "title": "Amount Original"
          },
          "amount_rub": {
            "type": "string",
            "title": "Amount Rub"
          },
          "source_system": {
            "type": "string",
            "title": "Source System"
          },
          "is_manual_entry": {
            "type": "boolean",
            "title": "Is Manual Entry"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "transaction_date",
          "account_id",
          "amount_original",
          "amount_rub",
          "source_system",
          "is_manual_entry",
          "created_at"
        ],
        "title": "TransactionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  }
}